pyproject.toml
[tool.poetry]
name = "lc"
version = "0.1.0"
description = ""
authors = ["jsb19 <jujhar.bedi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<3.12"
pandas = "<2.0.0"
python-dotenv = "^1.0.0"
langchain = ">=0.0.239,<=0.0.293"
openai = "0.27.5"
scikit-learn = "^1.3.0"
llamaapi = "^0.1.36"
streamlit-chatbox = "^1.1.9"
streamlit = "^1.26.0"
llama-index = "^0.8.33"
torch = [
  {url = "https://download.pytorch.org/whl/cpu/torch-2.1.0%2Bcpu-cp39-cp39-linux_x86_64.whl", "platform" = "linux"},
  {url = "https://download.pytorch.org/whl/cpu/torch-2.1.0%2Bcpu-cp39-cp39-win_amd64.whl", "platform" = "win32"},
  {url = "https://download.pytorch.org/whl/cpu/torch-2.1.0-cp39-none-macosx_10_9_x86_64.whl", "platform" = "darwin",  markers = "platform_machine == 'x86_64'" },
  {url = "https://download.pytorch.org/whl/cpu/torch-2.1.0-cp39-none-macosx_11_0_arm64.whl", "platform" = "darwin",  markers = "platform_machine == 'arm64'" }
]
transformers = "^4.33.3"
clip = "^0.2.0"
matplotlib = "^3.8.0"
opencv-python = "^4.8.1.78"
tensorflow = "^2.14.0"
openai-clip = "^1.0.1"
torchvision = "^0.16.0"
flask = "^3.0.0"
fastapi-cache2 = {version = "0.1.9", extras = ["redis"]}
redis = "^4.6.0"
neo4j = "^5.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
